+++ src/main/scala/com/deloitte/demoApp/MainApp.scala
@@ -1,6 +1,6 @@
 package com.deloitte.demoApp
 
-import com.deloitte.demoApp.operations.OperationsHandler
+import com.deloitte.demoApp.handler.OperationsHandler
 
 import scala.util.Try
 
@@ -12,6 +12,7 @@ object MainApp extends App {
   } else {
     Try {
       println(s"Operation not found for symbol << ${arguments(0)} >>")
+      operation.printHelp
     } getOrElse println("Error identifying operation")
   }
 }
